Initial concrete impl of ts serverSentEventGenerator
This can do cross-runtime parsing of signals, but I havent figured out
yet how to do the streaming cross-runtime, the current code is for node only.

I will probably figure this out tomorrow.
